+++ app/src/main/java/woowacourse/movie/helper/ImageViewBindingAdapter.kt
@@ -1,0 +1,12 @@
+package woowacourse.movie.helper
+
+import android.widget.ImageView
+import androidx.databinding.BindingAdapter
+
+@BindingAdapter("drawableResource")
+fun setDrawableResource(
+    view: ImageView,
+    drawableRes: Int,
+) {
+    view.setImageResource(drawableRes)
+}
